api/requirements.txt
flask
flask-cors
marshmallow
pydantic
pydantic-settings
transformers
torch
tokenizers
nltk
python-dotenv
requests
pytest
pytest-flask
black
flake8
mypy
gunicorn
structlog
pyPDF2
accelerate
dataset